+++ app/build.gradle.kts
@@ -127,7 +127,6 @@ dependencies {
     implementation(libs.kotlinx.serialization)
     implementation(libs.kotlinx.datetime)
     implementation(libs.kotlinx.serialization.retrofit)
-    implementation(libs.converter.gson.v290)
 
     /* Cronet and Coil */
     coreLibraryDesugaring(libs.desugaring)
