+++ app/src/main/java/com/github/libretube/constants/PreferenceKeys.kt
@@ -136,6 +136,7 @@ object PreferenceKeys {
     /**
      * Advanced
      */
+    const val AUTO_UPDATES = "auto_update"
     const val DATA_SAVER_MODE = "data_saver_mode_key"
     const val MAX_IMAGE_CACHE = "image_cache_size"
     const val RESET_SETTINGS = "reset_settings"
