+++ app/src/main/res/xml/advanced_settings.xml
@@ -4,6 +4,12 @@
 
     <PreferenceCategory app:title="@string/advanced">
 
+        <SwitchPreferenceCompat
+            android:defaultValue="false"
+            android:icon="@drawable/ic_download_filled"
+            app:key="auto_update"
+            app:title="@string/show_updates" />
+
         <ListPreference
             android:defaultValue="disabled"
             android:entries="@array/dataSaverModeOptions"
